Luke: Refactored `ingestion' -> `last_update'
